Add Debian Buster image doc
We're working on updating the Hypernode platform from Ubuntu Xenial to Debian Buster. We're publishing Docker images for Buster which can already be used for testing. New in Buster is that we'll also have images for specific configurations so a different version of MySQL than the default does not have to be manually installed.